pyproject.toml
[project]
name = "doc-intel-engine"
version = "0.1.0"
description = "Production-oriented document intelligence service with OCR and LLM classification."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4==4.14.3",
  "fastapi==0.135.1",
  "httpx==0.28.1",
  "markdown==3.10.2",
  "numpy==2.4.3",
  "openai==2.28.0",
  "paddleocr[doc-parser]==3.4.0",
  "pillow==12.1.1",
  "pydantic==2.12.5",
  "pydantic-settings==2.13.1",
  "pypdf==6.9.0",
  "pypdfium2==5.6.0",
  "python-docx==1.2.0",
  "python-jose==3.5.0",
  "python-multipart==0.0.22",
  "slowapi==0.1.9",
  "sqlalchemy==2.0.48",
  "structlog==25.5.0",
  "tenacity==9.1.4",
  "tiktoken==0.12.0",
  "uvicorn[standard]==0.41.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
